requirements.txt
flask
requests
python-telegram-bot==20.7
deriv-api